Changed Windowed to save settings
The Options were extended to save also the last set Width and Height of the window. The UpdateWindow function was modified to use this setting if it is set. When changing the view in options to windowed its reseted to a maximized window, by setting the default window size of 640x480.